requirements.txt
absl-py==2.1.0
altair==4.2.2
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
entrypoints==0.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
idna==3.7
IETK-Ret==0.1.1
imageio==2.34.1
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
networkx==3.3
numpy==1.26.2
oauthlib==3.2.2
opencv-contrib-python==4.5.4.58
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.2
pillow==10.2.0
protobuf==3.19.6
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydeck==0.9.0
Pygments==2.17.2
Pympler==1.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-image==0.23.2
scipy==1.13.0
six==1.16.0
smmap==5.0.1
streamlit==1.22.0
tenacity==8.2.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tifffile==2024.4.24
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.28.1
watchdog==4.0.0
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1